requirements.txt
### requirements.txt
fastapi
uvicorn[standard]
langchain
langchain_community
langchain-tavily
langgraph
langchain-openai
python-jose 
passlib[bcrypt]
python-multipart
streamlit
python-dotenv
pymongo
